python/requirements.txt
Flask
pyserial
keyboard
websockets